Improved Tape Design (1.1)
Previously, due to the way the tape rotated, whenever a tape rotation
happened the pointer would be pointing to the same element it was
pointing at before execution of a SPIN command. Now the pointer rotates
with the tape. This propogates back up as well, so if the most outward
level of tape rotates, then all the layers above it will also rotate.
This means that every time a SPIN command is executed, the pointer will
always be pointing at a new memory location.