Disabled automatic tape creation
It might be too much to deal with. Also included a program that sets [5]
to 1 or 0, if X > Y (user input) is true/false. I made a version that
uses neg and one that *does not* use neg. This is the basis for making
an X + Y that works regardless of X and Y's size.

Note that this does not work if X or Y are negative.