Debug features added
Added a flag in main.cpp, debugFeatures. When it's true, commands
starting with - can be used. These are debug commands meant for use
while testing programs and the language. Right now all that's
implemented is -dumpX, where X is a number. It will print the values of
the tape from [0] until X-1.

If debugFeatures is false and there are debug commands in the program it
will still run, but it will print an error.